+++ reactjs/src/lib/moment.ts
@@ -8,6 +8,7 @@ export const nowMoment = moment();
 
 export const fromNow = (date: Date) => moment(date).fromNow() || '';
 export const isClosed = (date: Date) => moment(date).diff(now, 'days') <= 0;
+export const isToday = (date: Date) => moment(date).isSame(moment(now), 'days');
 
 export const formatDatetime = (date: Date) =>
   moment(date).format('MM月DD日 HH:mm');
