+++ reactjs/src/web/components/molecules/challenges/ChallengePostController.tsx
@@ -1,21 +1,14 @@
 import * as React from 'react';
 import styled from 'styled-components';
-import moment from 'moment';
 
 import { useDocument } from 'react-firebase-hooks/firestore';
-import shortId from 'shortid';
 import { Button } from '@material-ui/core';
 import { withRouter } from 'react-router-dom';
-import firebase from '~/lib/firebase';
-
-import { postMessage } from '~/lib/discord.client.api';
-
-import rollbar from '~/lib/rollbar';
-import { getUserDashboardPath, withDomain } from '~/lib/url';
-import { getParticipantsUserId } from '~/lib/resource';
 
 import Error from '../../atoms/Error';
 import { brandSuccess, brandWhite, brandWarning } from '~/lib/theme';
+
+import { isDaysValid } from '~/lib/challenge';
 
 const StyledCenterContainer = styled.div`
   display: flex;
@@ -25,137 +18,16 @@ const StyledCenterContainer = styled.div`
 `;
 
 const ChallengePostController = (props: any) => {
-  const { userShortId, history } = props;
-  const { webhookURL, openedAt, closedAt, id } = props.challenge;
+  const { history, recordHandler, resetHandler, hide, participantsRef } = props;
 
-  const challengeId = id;
-  const resourceId = getParticipantsUserId(challengeId, userShortId);
+  const [value, loading, error] = useDocument(participantsRef);
 
-  const [value, loading, error] = useDocument(
-    firebase.firestore().doc(resourceId)
+  const writeRecord = recordHandler(
+    window.alert, // eslint-disable-line
+    history.push
   );
 
-  const now = new Date();
-  const isDaysValid = (days: number) => {
-    return days !== undefined && days !== null;
-  };
-
-  const writeRecord = (props: any) => {
-    const {
-      days,
-      score,
-      accDays,
-      maxDays,
-      pastDays,
-      histories,
-      displayName
-    } = props;
-
-    if (
-      histories.length > 0 &&
-      histories.filter((history: any) =>
-        moment(history.timestamp.toDate()).isSame(moment(now), 'days')
-      ).length !== 0
-    ) {
-      window.alert('記録の投稿は1日1回までです。'); // eslint-disable-line
-      return;
-    }
-
-    const tomorrow = !isDaysValid(days) ? 1 : days + 1;
-    const newPastDays = !isDaysValid(pastDays) ? 1 : pastDays + 1;
-    const newScore = score + 1;
-    const newAccDays = !isDaysValid(accDays) ? 1 : accDays + 1;
-    const newMaxDays = tomorrow > maxDays ? tomorrow : maxDays;
-
-    const newHistory = {
-      id: shortId.generate(),
-      timestamp: new Date(),
-      score: newScore,
-      days: tomorrow,
-      accDays: newAccDays,
-      pastDays: newPastDays,
-      diff: moment().diff(moment(openedAt.toDate()), 'days'),
-      type: 'RECORD'
-    };
-
-    const updateData: any = {
-      days: tomorrow,
-      score: newScore,
-      accDays: newAccDays,
-      pastDays: newPastDays,
-      maxDays: newMaxDays,
-      startedAt: now,
-      updatedAt: now,
-      histories: firebase.firestore.FieldValue.arrayUnion(newHistory)
-    };
-
-    if (!isDaysValid(days)) updateData.startedAt = now;
-
-    firebase
-      .firestore()
-      .doc(resourceId)
-      .update(updateData)
-      .then(() => {
-        const url = withDomain(getUserDashboardPath(challengeId, userShortId));
-        const message = `${displayName}さんが計${newAccDays}日達成しました！
-${url}`;
-        postMessage(webhookURL, message);
-      })
-      .then(() => {
-        window.alert('投稿が完了しました。'); // eslint-disable-line
-      })
-      .then(() => history.push(getUserDashboardPath(challengeId, userShortId)))
-      .then(() => window.location.reload()) // eslint-disable-line
-      .catch(error => rollbar.error(error));
-  };
-
-  const resetRecord = (props: any) => {
-    const { score, displayName, accDays, histories } = props;
-
-    const newScore = score - 3;
-
-    const lastHistory = histories[histories.length - 1];
-    const newAccDays =
-      moment(lastHistory.timestamp.toDate()).isSame(moment(now), 'days') &&
-      lastHistory.type === 'RECORD'
-        ? accDays - 1
-        : accDays;
-
-    const newHistory = {
-      id: shortId.generate(),
-      timestamp: new Date(),
-      score: newScore,
-      days: 0,
-      pastDays: 0,
-      accDays: newAccDays,
-      diff: moment().diff(moment(openedAt.toDate()), 'days'),
-      type: 'RESET'
-    };
-
-    const resetData = {
-      startedAt: null,
-      updatedAt: now,
-      days: 0,
-      pastDays: 0,
-      score: newScore,
-      accDays: newAccDays,
-      histories: firebase.firestore.FieldValue.arrayUnion(newHistory)
-    };
-
-    firebase
-      .firestore()
-      .doc(resourceId)
-      .update(resetData)
-      .then(() => {
-        const url = withDomain(getUserDashboardPath(challengeId, userShortId));
-        const message = `${displayName}さんがリセットしました。
-${url}`;
-        postMessage(webhookURL, message);
-      })
-      .then(() => history.push(getUserDashboardPath(challengeId, userShortId)))
-      .then(() => window.location.reload()) // eslint-disable-line
-      .catch(error => rollbar.error(error));
-  };
+  const resetRecord = resetHandler(history.push);
 
   const confirm = (props: any) => {
     const { days } = props;
@@ -169,54 +41,6 @@ ${url}`;
   };
 
   const data = value && value.data();
-  const hide = !(
-    moment(now).diff(moment(openedAt.toDate())) >= 0 &&
-    moment(now).diff(moment(closedAt.toDate())) < 0
-  );
-
-  // useEffect(() => {
-  //   if (
-  //     data !== undefined &&
-  //     data.updatedAt !== undefined &&
-  //     data.updatedAt !== null &&
-  //     moment(now).diff(
-  //       moment(new Date(data.updatedAt.toDate().setHours(0, 0, 0, 0))),
-  //       'days'
-  //     ) > 1 &&
-  //     moment(now).diff(moment(openedAt.toDate()), 'days') !== 0 &&
-  //     moment(now).diff(moment(closedAt.toDate()), 'days') <= 0
-  //   ) {
-  //     const newScore = data.score - 3;
-
-  //     const newHistory = {
-  //       id: data.histories.length + 1,
-  //       timestamp: new Date(),
-  //       score: newScore,
-  //       days: 0,
-  //       pastDays: 0,
-  //       diff: moment().diff(moment(openedAt.toDate()), 'days'),
-  //       type: 'RESET'
-  //     };
-
-  //     const resetData = {
-  //       startedAt: null,
-  //       updatedAt: now,
-  //       days: 0,
-  //       pastDays: 0,
-  //       score: newScore,
-  //       histories: firebase.firestore.FieldValue.arrayUnion(newHistory)
-  //     };
-
-  //     firebase
-  //       .firestore()
-  //       .doc(resourceId)
-  //       .update(resetData)
-  //       .then(
-  //         () =>
-  //           window.alert('1日以上記録がなかったため、記録をリセットしました') // eslint-disable-line
-  //       );
-  //   }
-  // }, [closedAt, data, now, openedAt, resourceId]);
 
   return (
     <StyledCenterContainer>
